bnbapp/airbnb_backend/requirements.txt
Django
Django==3.2
djangorestframework==3.12.4
django-cors-headers==3.10.0
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.1